+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 0.7.0 (the future)
 ==================
+### New Features
+- Add `plotly_plot_density` function: Plots interactive density plots of the given variables using kernel density estimation.
 
 ### Bug Fixes
 - Fix `rm_mutual_nas`: it now remembers the index of pandas.Series that were used as input. This fixed erroneous plots produced by seaborn functions which uses rm_mutual_nas.